mptcp: add mptcp reset option support
Git-commit: dc87efdb1a5cd46134a9d490480160e303bc6eef
Message-Id: <dc87efdb1a5cd46134a9d490480160e303bc6eef.1637073388.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.13-rc1

The MPTCP reset option allows to carry a mptcp-specific error code that
provides more information on the nature of a connection reset.

Reset option data received gets stored in the subflow context so it can
be sent to userspace via the 'subflow closed' netlink event.

When a subflow is closed, the desired error code that should be sent to
the peer is also placed in the subflow context structure.

If a reset is sent before subflow establishment could complete, e.g. on
HMAC failure during an MP_JOIN operation, the mptcp skb extension is
used to store the reset information.

Signed-off-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.13-rc1-mptcp-add-mptcp-reset-option-support.patch