KVM: x86/mmu: Rename goal_gfn to next_last_level_gfn
Git-commit: 1cd17c5c9b8a6c2d550c9dbc934dffa813fccfb6
References: bsn#19
Patch-mainline: v5.10.30

[ Upstream commit 74953d3530280dc53256054e1906f58d07bfba44 ]

The goal_gfn field in tdp_iter can be misleading as it implies that it
is the iterator's final goal. It is really a target for the lowest gfn
mapped by the leaf level SPTE the iterator will traverse towards. Change
the field's name to be more precise.

Signed-off-by: Ben Gardon <bgardon@google.com>
Message-Id: <20210202185734.1680553-13-bgardon@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.30-0083-KVM-x86-mmu-Rename-goal_gfn-to-next_last_level_gfn.patch